file-mode-api: fix classes and names to fit our vocabulary.
