+++ airbyte_cdk/sources/declarative/retrievers/file_uploader/__init__.py
@@ -1,15 +1,15 @@
-from .base_file_uploader import BaseFileUploader
-from .base_file_writer import BaseFileWriter
-from .connector_builder_file_uploader import ConnectorBuilderFileUploader
 from .file_uploader import FileUploader
 from .file_writer import FileWriter
+from .connector_builder_file_uploader import ConnectorBuilderFileUploader
+from .default_file_uploader import DefaultFileUploader
+from .local_file_system_file_writer import LocalFileSystemFileWriter
 from .noop_file_writer import NoopFileWriter
 
 __all__ = [
-    "FileUploader",
-    "FileWriter",
+    "DefaultFileUploader",
+    "LocalFileSystemFileWriter",
     "NoopFileWriter",
     "ConnectorBuilderFileUploader",
-    "BaseFileUploader",
-    "BaseFileWriter",
+    "FileUploader",
+    "FileWriter",
 ]
