+++ airbyte_cdk/sources/declarative/retrievers/file_uploader/file_writer.py
@@ -2,17 +2,18 @@
 # Copyright (c) 2025 Airbyte, Inc., all rights reserved.
 #
 
+from abc import ABC, abstractmethod
 from pathlib import Path
 
-from .base_file_writer import BaseFileWriter
 
+class FileWriter(ABC):
+    """
+    Base File writer class
+    """
 
-class FileWriter(BaseFileWriter):
+    @abstractmethod
     def write(self, file_path: Path, content: bytes) -> int:
         """
         Writes the file to the specified location
         """
-        with open(str(file_path), "wb") as f:
-            f.write(content)
-
-        return file_path.stat().st_size
+        ...
