+++ airbyte_cdk/sources/declarative/retrievers/file_uploader/local_file_system_file_writer.py
@@ -2,18 +2,17 @@
 # Copyright (c) 2025 Airbyte, Inc., all rights reserved.
 #
 
-from abc import ABC, abstractmethod
 from pathlib import Path
 
+from .file_writer import FileWriter
 
-class BaseFileWriter(ABC):
-    """
-    Base File writer class
-    """
 
-    @abstractmethod
+class LocalFileSystemFileWriter(FileWriter):
     def write(self, file_path: Path, content: bytes) -> int:
         """
         Writes the file to the specified location
         """
-        ...
+        with open(str(file_path), "wb") as f:
+            f.write(content)
+
+        return file_path.stat().st_size
